@@ -1,12 +1,6 @@
 <template>
   <div>
-    <h2>Welcome to the POS System</h2>
-    <p>Manage your sales, inventory, and customers easily.</p>
+    <h1>Welcome to POS System</h1>
+    <p>POS and Inventory Management</p>
   </div>
 </template>
-
-<script>
-export default {
-  name: 'Home'
-};
-</script>
